cryptohome: Check lockbox nvram when using TPM2.0 simulator
Let the lockbox-cache start just before cryptohomed because cryptohomed
is the only consumer of it, and guarant it started after trunksd
started.

This would prevent the lockbox cache appear after clear TPM2.0 simulator
ownership.

If the lockbox cache appear, cryptohome would consider the lockbox is
initialized. But cryptohome would have a check failed and crash when it
found the lockbox is initialized and TPM is not owned.

And the tast-test would try to check are the attributes still there
after cleared TPM ownership, without this CL, the attributes would
always be installed on the VM after first installed.

BUG=b:170785530
TEST=hwsec.CryptohomeKeysAndMountSanity hwsec.InstallAttributes

Cq-Depend: chromium:2470577, chromium:2505239
Change-Id: Iecce1e36b5b4ea45e0e6717308bd7c0b3fd6292e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2482225
Tested-by: joe Chou <yich@google.com>
Reviewed-by: Leo Lai <cylai@google.com>
Commit-Queue: joe Chou <yich@google.com>