mount-passthrough: Introduce new script to set up "Play files"
In ARC container, files under Android's /data/media/0 are accessible in
the Files app as "Play files". We are going to enable the same feature
in ARCVM. However, due to changes in the filesystem permissions of the
files owned by Android, the Files app cannot access these files in ARCVM
as-is.
We solve this problem by translating their filesystem permissions using
mount-passthrough. Note that this usage of mount-passthrough is
different from the previous ones; It has been used to let Android access
files owned by chronos. On the other hand, this time we are going to use
it to let Chrome (the Files app) access files owned by Android.
For this new usage of mount-passthrough, this CL adds several new
command line arguments to mount-passthrough, and introduces a new script
mount-passthrough-jailed-play to call mount-passthrough with an
appropriate setting.
When the mount-passthrough daemon is called from the new script, (1) it
runs with CAP_DAC_OVERRIDE so that it can access Android files, and (2)
it runs as Android's media_rw user and creates new directories with full
group access permission so that files created by the Files app can be
accessed by Android apps.

BUG=b:123669632
BUG=b:148298890
TEST=With crrev.com/c/2537511,
TEST=for each <package> in {arcvm-mount-media-dirs, mount-passthrough},
TEST=$ cros_workon --board=${BOARD} start <package>
TEST=$ emerge --board=${BOARD} <package>
TEST=$ cros deploy --board=${BOARD} <DUT> <package>
TEST=Start ARCVM, confirm that the "Play files" pane in the Files app is
TEST=populated by Android's /data/media/0.
TEST=Create a directory in the Files app's "Play files" pane, and copy a
TEST=file into the directory using the Files app. Confirm that the
TEST=copied file can be accessed by Android apps like Photos.
TEST=Select a directory in the Files app's "Play files" pane > Share
TEST=with Linux > Open with Terminal. Confirm that the directory is
TEST=successfully opened in Terminal.

Cq-Depend: chromium:2626688
Change-Id: Ic1d5ce0d63bb26de6c6cb13190686154a71006bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2483703
Tested-by: Youkichi Hosoi <youkichihosoi@chromium.org>
Commit-Queue: Youkichi Hosoi <youkichihosoi@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>