libchromeos-rs: Add read_dir
Add a function to read the contents of a directory.  Unlike
std::fs::read_dir, this takes an offset to allow skipping any entries
that we previously returned.  Similarly, the returned DirEntry includes
the offset that callers can use to pass to subsequent calls to read_dir.

This is mostly a port of the read_dir function from crosvm and we'll
switch crosvm over to use this one once it lands.

BUG=none
TEST=unit tests and 9p tests with later changes in this series

Cq-Depend: chromium:2494293
Change-Id: Ia40d6e9f160cb290f08633feac8721fcb4a63ef6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2494168
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Tested-by: Chirantan Ekbote <chirantan@chromium.org>
Commit-Queue: Chirantan Ekbote <chirantan@chromium.org>