system_api: Add request_id parameter for WebAuthn requests
This enables us to identify the browser window that initiated the
request. U2fd will pass this request_id to the authentication
dialog UI, so that the authentication dialog can be modal to the
browser window that initiated the flow.

BUG=none
TEST=none

Change-Id: I8c049630d593e90af33929f60e4d6995c314ce76
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2487542
Tested-by: Yicheng Li <yichengli@chromium.org>
Auto-Submit: Yicheng Li <yichengli@chromium.org>
Commit-Queue: Leo Lai <cylai@google.com>
Reviewed-by: Leo Lai <cylai@google.com>