vm_tools: sommelier: Don't send axis events with 0 units of scroll
In CL:2315787 we started buffering axis and axis_discrete events until
the next frame event, which caused janky scrolling behavior in some
apps.

It's not entierly clear why this happened, but it's probably because
we started sending axis events even when no scrolling on either axis
had happened. Fix this by not sending an event if the scroll amount
would be 0.

BUG=chromium:1126610
TEST=Ran stellarium and tried zooming in and out

Change-Id: I4a3fbaf4df45544c483d58856c6fccbaecc609d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2490406
Tested-by: Fergus Dall <sidereal@google.com>
Reviewed-by: Nic Hollingum <hollingum@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>