vm_tools: p9: Stop using path-based operations
The main change in this CL is to use O_PATH fds instead of keeping a
separate path for each fid that we had open.  This helps avoid some race
conditions inherent to dealing directly with paths.  This is also needed
in order to support ascii casefolding in a reasonable way.

The rest of the CL deals with the fallout of making this change.  The
biggest one is that we lose support for remove and rename but this is ok
because we only support the 9P2000.L version of the protocol, which
instead uses unlink_at and rename_at, respectively.

BUG=b:162388795
TEST=`tast run $DUT crostini.Share*.artifact`

Cq-Depend: chromium:2494294, chromium:2494328
Change-Id: Ia95bf14b0abe4e39be0098ef2523770826e4ab5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2494169
Commit-Queue: Chirantan Ekbote <chirantan@chromium.org>
Tested-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>