cryptohome: Add tmpfs proxy fakes to FakePlatform
And use them in one place as a proof of correctness. Apparently some
places of the tests already have sudo fakes, but they are very limited
and intermixed with the mocks in a very inconsistent way. Later we will
unify all of the cases to be 100% consistent and use FakePlatform.

BUG=chromium:1141301
TEST=unittest

Change-Id: Ic1e5c10ccd9addf74646d9d04da8cd3e77e60ab5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2497362
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Tested-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Daniil Lunev <dlunev@chromium.org>