cryptohome: add more fs related calls to FakePlatform
After this addition, we get enough mass to start converting some tests.
The first target is to convert tests in homedirs, which is being done in
the following CL.
Some of the added functions are not FS proxies, but pure in-memory fakes
because of the lack of permissions of the test-runner - we can consider
running it as root later.
Some of the unittests are patched up because they started failing after
the redirection to fakes - they were using obscure mock setup methods.

BUG=chromium:1141301
TEST=unittest

Change-Id: I1eec914667c7692771879f3083d30f483ae700f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2498161
Tested-by: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Sergei Datsenko <dats@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Commit-Queue: Daniil Lunev <dlunev@chromium.org>