power: Add GetThermalState API
Although the thermal change event has already been implemented in
Thermal Hinting API and an event will be sent with the state when
initialized, it is still impossible for Chromium to get the initial
state if it missed the first event sent. So in this patch, a new
API called GetThermalState is implemented, making it possible to
read the current thermal state any time the browser wishes to.

BUG=chromium:1081959
TEST=FEATURES=test emerge-atlas power_manager

Change-Id: I9de187631f59de78e63db2dbf62d77300587ba61
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2495080
Tested-by: Jiahe Zhang <jiahe.zhang@intel.com>
Commit-Queue: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Auto-Submit: Jiahe Zhang <jiahe.zhang@intel.com>