permission_broker: allow detaching usb-serial converters
Allow detaching devices presented as usb-serial interface in order to be
able to use them in chrome.usb, WebUSB and VMs (Crostini, ARC).

If a USB device is composed only of unclaimed interfaces and usb-serial
interfaces, now returns ALLOW_WITH_DETACH for the full device, but if
the devices also has other claimed interfaces still deny it in order to
be conservative as they might be some kind of privileged interfaces to
the device internal.

An interface is considered a 'usb-serial' one iff it has a UDEV child in
the 'usb-serial' subsystem as this means the interface is handled by a
kernel driver deriving from usb_serial (e.g ftdi_sio, ch341, pl2303,
usb_serial_simple). This does NOT include the modems CDC-ACM interfaces.

BUG=chromium:987377
TEST=manual, plug plenty of USB devices (FTDI converters, PL2303 converters,
Arduino, Servos, LTE Modems, ...) and stare at the permission_broker
logs to verify we are detecting the right thing and doing the right
choice.

Change-Id: I3284acab2e59070eb5d9e4f1ec9c158cd638f0a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2502688
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>