shill: cellular: use operator name provided by MM if available
This changes the logic for deciding the name for operator
info, so that values provided by MM on the SIM (home) and
Modem3gpp (serving) will be used if available. It also uses
the name of the home operator instead of the serving operator
if we are not roaming and the home operator name is available.

BUG=b:171159326,b:171284831
TEST=unit tests, deploy to device with MM reporting "PetalNet"
  operator on SIM object and check that this is reproduced in
  the UI

Change-Id: If194babe675badb92b0ea0ce593750165d722f3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2503451
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Commit-Queue: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Pavan Holla <pholla@google.com>