oobe_config: Switch to rollback powerwash
The data save mechanism for rollback will switch to a rollback powerwash
once the data was saved successfully.

Currently, this change has no effect as the update_engine already
writes "rollback" into the powerwash file.

There will be a second CL to make the update_engine schedule a full
powerwash even for rollback images. Then we will only schedule a
rollback powerwash once we can be sure the data was saved.
This will avoid devices being stuck in a state where all data was
deleted but the TPM was not reset.

BUG=chromium:1141802
TEST=manually tested failed and successful rollback

Change-Id: Ic0927d4a4d9e854175a27068c3cc63a6293a103d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2504302
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Reviewed-by: Miriam Polzer <mpolzer@google.com>
Commit-Queue: Miriam Polzer <mpolzer@google.com>
Tested-by: Miriam Polzer <mpolzer@google.com>