typecd: Add interface to control mode entry / exit
Create an interface which is used by typecd to check whether a device
supports mode entry and also triggers entry/exit into a particular mode.

The "prod" implementation of this consists of stubs for the time being.
This can be fleshed out with an actual implementation later.

Also add a simple unit test that utilizes a Mock with expectations to
make sure that the RunModeEntry code is correctly checking for mode
entry support and bailing if that is not available.

BUG=b:168653671
TEST=P2_TEST_FILTER="*.*" cros_workon_make --board=volteer --test
typecd passes

Change-Id: If5abb36a068cb51f4fec5389170cb545cdf6767b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2503837
Tested-by: Prashant Malani <pmalani@chromium.org>
Commit-Queue: Prashant Malani <pmalani@chromium.org>
Reviewed-by: Benson Leung <bleung@google.com>