shill: Remove a ref count loop and add an error message
This removes a confusing and potentially problematic
DeviceRefPtr from TrafficMonitor (which is owned by Device,
and normally gets destroyed at shutdown, but is not clearly
guaranteed to be).

This also adds some logging in Manager::Stop if unexpected
references to Device or Service remain.

BUG=b:171565140
TEST=unit tests

Change-Id: I58f4c925d6f52ee927d603a55798fe5e7107411c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2511045
Tested-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Steven Bennetts <stevenjb@chromium.org>