modemfwd: make the carrier firmware check more robust
In the firmware manifest, a given carrier firmware might be used for
multiple carrier UUIDs (e.g. a carrier and its MVNO).
In this case, when asked for the flashed carrier firmware version and UUID,
the l850gl-helper might return an arbitrary UUID (which is hardcoded on
their side) for this firmware rather than current real one.
Detect this corner case to avoid endlessly re-flashing the carrier
firmware by comparing the carrier firmware file for the flashed UUID
with the file for the detected UUID (rather than comparing just the
UUIDs).

BUG=b:171367064
TEST=unit-tests: added SkipCarrierWithTwoUuidSameFirmware test case,
fails before the change and pass with it.
TEST=manual testing: put 2 different UUID in the
firmware_manifest and make the l850gl-helper returns the wrong one (not
matching the network one) and verify that we are not doing useless
updates.

Change-Id: I6aab9535f0c7dee43844c4477138e850853e3aec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2507711
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>