shill: enable IPv6 per APN in mobile operator DB
This change allows us to enable IPv6 on a per-APN basis
to ensure we don't have regressions on carriers we are
unable to test.

In the past, IPv6 was totally disabled on cellular and we
would always request the default IP type, which happened
to be IPv4. This could change in ModemManager so we are now
supplyng ip-type=ipv4 for all connect requests by default.

The Cellular Device no longer disables IPv6 by default.

The mobile operator DB schema has been extended so we can
now specify that a dual-stack or IPv6 only connection should
be requested when using that APN.

BUG=chromium:726815,b:162701578
TEST=unit tests, manually add an ip_type to an operator in
  the MODB and ensure that the relevant IP type is requested
  at connect time

Change-Id: I8d87797fd62138faaa462502c06b2c25522ca567
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2508389
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>