cryptohome: Move keyset loading code from mount.cc
Now the whole corpora of the vault keyset retrieval logic is a single
call to HomeDirs::LoadUnwrappedKeyset. This will allow moving it easily
out of mount in the following cl.
The unittest for DecryptVaultKeyset present in mount were effectively
testing the "if" in ShouldReSave, and the Tpm/Crypto behaviour.
The ShouldReSave is now explicitly tested, and the rest is assumed to be
tested in the relevant components.

BUG=chromium:1124561
TEST=unittest

Change-Id: Ib9262804c63450c5579e8d1d6f151b62f1631699
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2507597
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Tested-by: Greg Kerr <kerrnel@chromium.org>
Tested-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Daniil Lunev <dlunev@chromium.org>