arc: Add arcvm-post-login-services.conf
Chrome will use this job to control ARCVM-related jobs.
After this CL, adding a new ARCVM-related job can be done without
submitting a Chromium CL which cannot be tested with Chromium OS CQ.

Starting/stopping this job waits for all dependent jobs.
The "4.5 Job Lifecycle" section of the Upstart cookbook document says
jobs started with "on starting" conditions are in the "started" state
when the "starting" event of the meta job completes.
The same goes for "on stopping" and post-stop.
http://upstart.ubuntu.com/cookbook/#job-lifecycle

BUG=b:171408795
TEST=Launch Play Store with and without CL:2517269

Change-Id: I0c34ed56c28e566b5fcbda49dd60fd1c5d3cbfe4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2517272
Commit-Queue: Ryo Hashimoto <hashimoto@chromium.org>
Tested-by: Ryo Hashimoto <hashimoto@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>