oobe_config: Clear TPM on restore fail
If restore fails in such a way that we would not be able to restore
install attributes, clear the TPM.

In the first stage, the whole file is decrypted and moved to
/var/lib/oobe_config_restore. If that fails we need to reset.

In the second stage, all files are copied back to their original
location. If install attributes is not copied back, reset.

BUG=chromium:1141802
TEST=manually tested failed and successful rollback

Change-Id: I455a45f6a1e8570b9da0e6aaa2074a9d2e7168e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2526240
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Miriam Polzer <mpolzer@google.com>
Tested-by: Miriam Polzer <mpolzer@google.com>
Commit-Queue: Miriam Polzer <mpolzer@google.com>