u2fd: Ask cryptohome for WebAuthn secret and use hash in u2f_generate
On user login, obtain the auth-time WebAuthn secret from cryptohomed,
and keep the hash of it in memory. In MakeCredential, send the hash of
the secret to cr50 because it's used in KH generation.

When auth-time user secret completes migration, we will make u2fd obtain
the unhashed secret upon user verification, and send that in u2f_sign.

BUG=b:144861739
TEST=On Nocturne, verified that u2fd obtains the secret upon login by
     adding logging.

Change-Id: Ifc5942d83d7fb5ac528077c9423bfd62a63233a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2527756
Commit-Queue: Yicheng Li <yichengli@chromium.org>
Tested-by: Yicheng Li <yichengli@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>