shill: cancel traffic counter callback to release Service refs
Have the traffic counter callback we pass to the patchpanel client code
call another callback, and cancel this second callback upon the
destruction of the Device object. We do this to release the
ServiceRefPtrs that the callback holds.

We can't directly cancel the callback we pass to the patchpanel client
code since it expects a OnceCallback (and we don't have access to
base::CancelableOnceCallback), so we work around this with another
callback.

BUG=b:171565140
TEST=unit tests
TEST=network.DefaultProfile no longer crashes

Change-Id: I019133520b5b4b432fc2d0b8232a2b158244168d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2518398
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>