debugd: cups_tool: Add foomatic command verification
Verify that the FoomaticRIPCommand in PPD files is valid before adding
the printer. It is better for a printer to fail to install than for it
to not work properly and be unable to print.

Remove root_mount_ns from RunAsUser() - the root mount should be
accessible even without this.

BUG=none
TEST=tast run localhost:9222 printer.AddGenericPrinter

Change-Id: Ie27c519ce176ceed3ae9f538c9bf49636a57e0bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2524360
Tested-by: Pranav Batra <batrapranav@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Pranav Batra <batrapranav@chromium.org>