cryptohome: tpm_manager: change the name of D-Bus interface.
Making the name to be consistent with the D-Bus service name. Note that
we leave the variable name of the proxy class as it is because renaming
it to just `tpm_manager_` might cause confusions, e.g., it is misleading
that the proxy doesn't implement all methods by tpm manager.

BUG=b:172748618
TEST=hwsec.Login
TEST=hwsec.Attestation

Change-Id: I205fb503f73ab351c99cc354081967241c108803
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2526001
Reviewed-by: joe Chou <yich@google.com>
Tested-by: Leo Lai <cylai@google.com>
Commit-Queue: Leo Lai <cylai@google.com>