vm_tools: p9: Allow setting attributes on directories
When setting attributes we always opened the fd with O_RDWR. However,
this doesn't work for directories, which can only be opened with
O_RDONLY so setting attributes on directories would fail with EISDIR.

Check the file type before opening the fd and use the appropriate flags.
Also add some tests to make sure this doesn't break in the future.

BUG=b:172019337
TEST=unit tests

Change-Id: If899568d41bbb489bda1588ea6eedf508a71d7a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2526006
Tested-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Chirantan Ekbote <chirantan@chromium.org>