power: Add init script to configure kernel thermal zones
The script is currently limited to configure the governor of given
thermal zones. The Linux kernel only allows to specify a default
governor, the use of a different governor can only be configured
from userspace.

The optional configuration file lists the names of the thermal zones
and the governor to be used:

  THERMAL_ZONES="gpuss0-thermal gpuss1-thermal"
  THERMAL_GOVERNOR="step_wise"

A more sophisticated format can be used if other features are needed
in the future.

BUG=b:171244475
TEST=on trogdor:
  governor for GPU thermal zones is changed with the above configuration

Change-Id: Ic7f89d88b5f1b46a40e77ccaf31aa85e1a545b2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2535977
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>