system-proxy: Build //net/ntlm code
This CL adapts and builds the //net/ntlm code and tests with the
following changes:
1. Replace boringssl with openssl. boringssl is not available on the
platform.
2. Remove reference to the "net_string_util.h" file because it requires
the third_party library icu to build. Instead, the conversion is made
using base::ToUpperASCII.
3. For the test data, explicitly convert hex codes to char. On the
platform, the hex codes are converted to unsigned chars by default,
causing compilation issues when the value is assigned to a char.
4. Redefine the NET_EXPORT and NET_EXPORT_PRIVATE macros to do nothing.

BUG=b:173097096
TEST=P2_TEST_FILTER="Ntlm*" cros_workon_make
    --board=eve system-proxy --test

Change-Id: I8c09a382e2542db2cf8e329d26fb7426621f2014
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2532006
Tested-by: Andreea-Elena Costinas <acostinas@google.com>
Commit-Queue: Andreea-Elena Costinas <acostinas@google.com>
Reviewed-by: Pavol Marko <pmarko@chromium.org>