diagnostics: make CPU stress routine parameters optional
Now that we have good defaults configured for the CPU stress, CPU
cache, floating point accuracy and prime search routines, we can
make their parameters optional. This will make the routines easier
to use, because clients of cros_healthd don't need to know what
reasonable values are for the routines.

Also make the prime search max_num parameter configurable from
cros_config - this doesn't need to be input to new prime search
routines on a regular basis, but if the OEM who wrote the routine
wants to configure it, we leave them that option.

BUG=b:173062802
TEST=USE=coverage FEATURES=test emerge-nami diagnostics

Change-Id: Ie2e18eed9fcb0a53fa0208a4f2eb9f83362593c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2537116
Tested-by: Paul Moy <pmoy@chromium.org>
Commit-Queue: Paul Moy <pmoy@chromium.org>
Reviewed-by: Trent Begin <tbegin@chromium.org>