u2fd: Persist WebAuthn auth-time secret hash to disk
U2fd obtains the auth-time secret from cryptohome on login, and keeps
the hash of it in memory. This hash is needed in MakeCredential.
If u2fd crashes, the hash in memory will be lost.

This change does two things:
1. On successfully obtaining the auth-time secret, we persist a copy
of the hash to daemon-store.
2. If we crash during a user session, we load the hash of auth-time
secret from disk, and also load the WebAuthn records which are needed
in GetAssertion.

BUG=b:144861739
TEST=On nocturne, verify that MakeCredential and GetAssertion both
     work after "restart u2fd".
TEST=Verify that the secret hash read back from disk is the same as
     the hash obtained at login, by logging.

Change-Id: Ie40c2e97bc1527f55831e79c7d2a7e22fd1fd498
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2538572
Commit-Queue: Yicheng Li <yichengli@chromium.org>
Tested-by: Yicheng Li <yichengli@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>