attestation: certificateion also returns public key.
The callers often query the public key after the attestation flow, and
it requires extra calls to attestation service, which isn't as efficient
as including the public key of interest in the reply directly.

BUG=b:173293956
TEST=unittests.
TEST=attestation_client get_cert --label=testlabel twice; the first run
checks the correctness for a new certificate, the the second run for an
existing one.

Change-Id: I13eb691fadcc20fdc0b1b4b7cbd44a2b20b837ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2539214
Commit-Queue: Leo Lai <cylai@google.com>
Tested-by: Leo Lai <cylai@google.com>
Reviewed-by: joe Chou <yich@google.com>