cryptohome: Cleanup HomeDirs initialization path
Presently, the HomeDirs object takes a weird place in the dependency
chain: it is a dependant of service class, but it provides some info to
it and instantiates the directory and file structure for the rest of the
cryptohomed to function properly. On top of that, it has pretty messy
initialization path, where part of the dependencies is passed
explicitly, part is generated under the hood, and part is passed through
side channels (especially in tests). This CL unconvolutes the hierarchy
and initialization path, making the global state to be generated
by a separate functions and all the dependencies to be explicitly passed
to HomeDirs constructor.

BUG=chromium:1124561
TEST=unittest
TEST=tast run <DUT> hwsec.*

Change-Id: Ice080fcdd2878de21b2823545d19c69afdbc6b6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2537827
Commit-Queue: Daniil Lunev <dlunev@chromium.org>
Tested-by: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>