libbrillo: Use OPENSSL_cleanse instead of unoptimized memset
In general, we want to use a single library for all secure operations.
Currently that is OpenSSL.

As a benefit, OPENSSL_cleanse is as effective in practice as the
volatile data pointer implementation, but significantly faster:
https://www.usenix.org/system/files/conference/usenixsecurity17/sec17-yang.pdf

This also renames SecureMemset to SecureClear to better match the
proposed standard in
http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2020/p1315r5.html.

As a side effect, the function signature also changes, but no users were
using the return value or setting the value to anything other than zero.

BUG=b:169602629
TEST=FEATURES="test" emerge-amd64-generic libbrillo

Change-Id: Ica6cb6542b17fafadb7bc00a12f9b68f55c96490
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2365643
Tested-by: Tom Hughes <tomhughes@chromium.org>
Reviewed-by: John Riordan <jrio@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Commit-Queue: Tom Hughes <tomhughes@chromium.org>