crash: Add random 5-digit number to filenames.
If two crashes occur in the same second from the same process (e.g. a
chrome DumpWithoutCrashing followed by a real crash, or two selinux
violations from the same process), the second crash will not be saved
because of a filename collision.

To mitigate this, add a random 5-digit number to filenames so that the
odds of a collision are far smaller. Assuming there are at most 10
crashes from the same process in the same second, the odds of a
collision are:

1 - 100,000!/(100,000^10 * (100,000 - 10)!)
  = 1 - 0.99955 = 0.00045, or roughly 1/2,222.

With only two such crashes, odds of a collision are 1/100,000.

(calculated using the equation from
https://en.wikipedia.org/wiki/Birthday_problem)

BUG=chromium:1149605
TEST=tast run $DUT crash.* platform.*Crash* ui.ChromeCrash* \
     platform.MountFailure arc.AppCrash* crostini.CrashReporter* \
     crostini.FsCorruption*
TEST=test_that $DUT logging_CrashSender logging_UserCrash \
     logging_GenerateCrashFiles.{RENDER,CHROME,KERNEL,PROCESS}

Cq-Depend: chromium:2543491,chromium:2545138,chromium:2545569
Change-Id: Ibfa57ec89b3038850ce25fa57e097ee728314b4f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2543117
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>