lorgnette: Add a bunch of temporary logging
Tast tests on betty are timing out on release builder runs.  The symptom
is that tast makes a dbus call to lorgnette, then times out a minute or
two later with no further log messages.  The failures can't be
reproduced on a local betty VM, so add a bunch of temporary logging
around the code paths that seem most likely to have potential delays or
hangs.  This will be reverted once we track down the cause of tast
failures.

BUG=b:159484077
TEST=tast run betty scanner.*

Change-Id: I1f5fd873cd2375c20421214049699d0ed848cde8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2542589
Reviewed-by: Fletcher Woodruff <fletcherw@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>