shill: cellular: clear the 'attach APN' when not present
If the APN database is not containing any APN with the property
'is_attach_apn' set for the current SIM (MNC), send instead empty
settings for the APN. This will clear the 'attach APN' configuration
on the modem side.

The description of the MBIM_CID_MS_LTE_ATTACH_CONFIG extension specifies
explicitly that the modem should the 'attach APN' configuration when it
detects a SIM Swap, but this is not the case for at least one of our
modems (L850-gl). So this is a workaround to avoid sending a wrong
'attach APN' to a carrier after a SIM swap.

BUG=b:160812573
TEST=manual, switch between a SIM having an Attach APN defined in the
database to one without it and verify that we register properly to the
network in both cases.

Change-Id: I89344fd3e7e8fbe4213111590cfd4856d7db78b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2543147
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>