sepolicy: Give kernel access to device directories
When init first loads the selinux policy, it relabels /dev while still
running in the kernel domain.  The kernel has permission to access and
relabel device files, but it generates avc denials when attempting to do
the same for labeled subdirectories in /dev.

Since the kernel is a permissive domain, these denials don't actually
prevent anything.  We already allow the kernel to relabel the device
files, so it makes sense to silence the log messages by explicitly
allowing permissions on subdirectories as well.

BUG=b:160243294
TEST=check dmesg for devtmpfs denials

Change-Id: I183daee697f5796d9f43aeb7994e8c03c75cb67b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2548461
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>