vm_tools: maitred: wait for disk expansion in resize
When expanding the disk, the host first resizes the disk image, then
calls maitred to resize the filesystem.  However, the notification from
host to guest of a disk resize is asynchronous and does not have a way
for the guest to signal back to the host that it has been recognized, so
it's possible for the filesystem resize to start before the guest kernel
has updated the block device size, resulting in flaky behavior when the
resize command runs first.

To avoid this, make the maitred ResizeFilesystem call wait until the
block device is at least as large as the requested size; this will have
no effect when shrinking (since the block device will always be larger),
and it will account for the race between the disk resize notification
and the filesystem resize.

BUG=chromium:1149634
TEST=tast run crostini.ResizeOk.download_buster

Change-Id: Id5792fc08141792992131b8fde91ce44633c8727
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2548563
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Reviewed-by: David Munro <davidmunro@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Tested-by: Daniel Verkamp <dverkamp@chromium.org>