libchromeos-rs: Fix clippy presubmit errors.
This removes the nightly feature because it was causing the presubmit
script to fail. The presubmit script calls
"cargo clippy --all-features"
This enables the nightly feature on a stable build and fails as a
result. Since the nightly feature doesn't seem to be used, remove it.

Once clippy was fixed for presubmit checks, clippy errors were found in
src/sync, so this includes fixes for those.

BUG=None
TEST=cargo clippy --all-features

Change-Id: I679a29baaafcc1c0310293cfe074c5aa74b84ca4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2551381
Tested-by: Allen Webb <allenwebb@google.com>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Commit-Queue: Allen Webb <allenwebb@google.com>