patchpanel: simplify traffic accounting iptables chains
This patch reduces the number of traffic accounting chains set in
iptables for a interface to only two: 1 for egress traffic, and 1 for
ingress traffic.

Because sources are matched on the fwmark source tag it is possible to
jump to the the same accounting rule for any egress traffic or any
ingress traffic that has been correctly marked. The only drawback is
that more rules have to be traversed.

BUG=b:160112868
BUG=b:161060333
TEST=Unit tests. Checked GetTrafficCounters still work with:
  $ dbus-send --system --dest=org.chromium.PatchPanel --print-reply \
    /org/chromium/PatchPanel org.chromium.PatchPanel.GetTrafficCounters

Change-Id: Ib2f74627bbf0460bc7c88dd462b737100b1b0672
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2562979
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Tested-by: Hugo Benichi <hugobenichi@google.com>