ippusb_manager: Accept all pending connections
ippusb_manager is normally started by upstart-socket-bridge when a
connection arrives on /run/ippusb/ippusb_manager.sock.  It accepts and
processes a single connection from the passed-in fd, then exits.
Because upstart-socket-bridge uses edge-triggered epoll, if a second
connection arrives before the first has been handled, the listening
socket never transitions back to the not-ready state and
upstart-socket-bridge never starts another copy.  This leaves the
IPP-USB system invisibly wedged until upstart-socket-bridge is
restarted.

To fix this, convert the single connection handler in ippusb_manager to
a loop that repeatedly accepts and handles new connections until the
passed-in fd is no longer ready for reading.  Because the socket should
be fully drained at the end of this loop, upstart-socket-manager will
see a new edge and can trigger a new socket event if a connection
arrives during the small window between the end of the loop and the
program exit.

BUG=b:159484077
TEST=tast run betty printer.* scanner.* in a loop for 12 hours

Disallow-Recycled-Builds: test-failures
Change-Id: I421e64c44a77417f56a79ed1d3641ba26836be4a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2551765
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Fletcher Woodruff <fletcherw@chromium.org>