fingerprint_study: Fix/hack thread gracefully term
When the user refreshes the page, the WebSocket connection is
supposed to close. On close, the signal is sent to have the
worker thread to gracefully terminate and "join" the main
thread that is managing the closure.
What really happens is that the worker thread never terminates
and the join operation times out after 10 seconds.

Unfortunately, self.client_terminated doesn't actually seem to reflect
this closing state, but self.server_terminated does.

To be honest, I suspect there are more edge cases to be found
here, but this fix is necessary for smooth operation.

BUG=b:144061947, b:173829819
TEST=Checked on host by refreshing the page

Change-Id: Ia14625f03a288441dcc5eeef6ca5eddbbf29cc9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2553008
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Tested-by: Craig Hesling <hesling@chromium.org>
Commit-Queue: Craig Hesling <hesling@chromium.org>