typecd: Add support to parse partner number of altmodes from sysfs.
Parse the number of alternate modes for a partner from the corresponding
sysfs attribute file. Also add a unit test to ensure that the value is
parsed and stored correctly.

While we are here, re-name one of the functions
UpdateAltModesFromSysfs() and re-purpose it to instead only parse the PD
Identity VDOs and the number of altmodes sysfs attributes (since the
altmodes get added by dedicated udev events themselves).

BUG=b:174057058
TEST=Run the typecd unit tests. Also monitor the logs of typecd on
volteer and ensure that the num_altmodes attribute gets parsed.

Change-Id: I481537771cab5937a7b2da9f3b9c3e11403e4f04
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2555680
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Commit-Queue: Prashant Malani <pmalani@chromium.org>
Tested-by: Prashant Malani <pmalani@chromium.org>